Add files via upload
This model is accurate!!!!. this model outputs the same guesses for the python script on its first iteration. So we have a completely working inference machine. now the next step would be developing a unfolded and pipelined function to replace the for loop and training the iteration using a golden list/ answer key to disassociate any in correct guesses with the incorrectly applied Test HV